requirement.txt
pypdf
langchain
torch
accelerate
bitsandbytes
ctransformers
sentence_transformers
chromadb
chainlit
huggingface_hub
python-dotenv